requirements.txt
asgiref==3.5.0
boto3==1.22.3
botocore==1.25.3
certifi==2021.10.8
charset-normalizer==2.0.11
# Django==4.0.2
django-js-asset==2.0.0
djangorestframework==3.13.1
idna==3.3
install==1.3.5
jmespath==1.0.0
psycopg2-binary==2.8.6
Pillow==9.0.0
python-dateutil==2.8.2
pytz==2021.3
gunicorn==20.1.0
requests==2.27.1
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
Unidecode==1.3.2
urllib3==1.26.8